refactor: move database-entity-related functions from `utils` to `backend-utils` (#505)
