+++ pages/api/atlases/[atlasId]/component-atlases.ts
@@ -1,5 +1,5 @@
+import { dbComponentAtlasToApiComponentAtlas } from "../../../../app/apis/catalog/hca-atlas-tracker/common/backend-utils";
 import { ROLE_GROUP } from "../../../../app/apis/catalog/hca-atlas-tracker/common/constants";
-import { dbComponentAtlasToApiComponentAtlas } from "../../../../app/apis/catalog/hca-atlas-tracker/common/utils";
 import { METHOD } from "../../../../app/common/entities";
 import { getAtlasComponentAtlases } from "../../../../app/services/component-atlases";
 import { handler, method, role } from "../../../../app/utils/api-handler";
