+++ pages/api/atlases/[atlasId]/component-atlases/[componentAtlasId].ts
@@ -1,4 +1,4 @@
-import { dbComponentAtlasToApiComponentAtlas } from "app/apis/catalog/hca-atlas-tracker/common/utils";
+import { dbComponentAtlasToApiComponentAtlas } from "app/apis/catalog/hca-atlas-tracker/common/backend-utils";
 import { ROLE_GROUP } from "../../../../../app/apis/catalog/hca-atlas-tracker/common/constants";
 import { ROLE } from "../../../../../app/apis/catalog/hca-atlas-tracker/common/entities";
 import { componentAtlasEditSchema } from "../../../../../app/apis/catalog/hca-atlas-tracker/common/schema";
