+++ pages/api/atlases/[atlasId]/source-studies/[sourceStudyId]/source-datasets.ts
@@ -1,5 +1,5 @@
+import { dbSourceDatasetToApiSourceDataset } from "../../../../../../app/apis/catalog/hca-atlas-tracker/common/backend-utils";
 import { ROLE_GROUP } from "../../../../../../app/apis/catalog/hca-atlas-tracker/common/constants";
-import { dbSourceDatasetToApiSourceDataset } from "../../../../../../app/apis/catalog/hca-atlas-tracker/common/utils";
 import { METHOD } from "../../../../../../app/common/entities";
 import { getSourceStudyDatasets } from "../../../../../../app/services/source-datasets";
 import { handler, method, role } from "../../../../../../app/utils/api-handler";
