+++ pages/api/comments/[threadId]/comments.ts
@@ -1,5 +1,5 @@
+import { dbCommentToApiComment } from "../../../../app/apis/catalog/hca-atlas-tracker/common/backend-utils";
 import { newCommentSchema } from "../../../../app/apis/catalog/hca-atlas-tracker/common/schema";
-import { dbCommentToApiComment } from "../../../../app/apis/catalog/hca-atlas-tracker/common/utils";
 import { METHOD } from "../../../../app/common/entities";
 import {
   createComment,
