+++ pages/api/tasks/[validationId]/comment.ts
@@ -1,6 +1,6 @@
+import { dbCommentToApiComment } from "../../../../app/apis/catalog/hca-atlas-tracker/common/backend-utils";
 import { ROLE } from "../../../../app/apis/catalog/hca-atlas-tracker/common/entities";
 import { newCommentThreadSchema } from "../../../../app/apis/catalog/hca-atlas-tracker/common/schema";
-import { dbCommentToApiComment } from "../../../../app/apis/catalog/hca-atlas-tracker/common/utils";
 import { METHOD } from "../../../../app/common/entities";
 import {
   createValidationComment,
