+++ pages/api/tasks/completion-dates.ts
@@ -1,6 +1,6 @@
+import { dbValidationToApiValidation } from "app/apis/catalog/hca-atlas-tracker/common/backend-utils";
 import { ROLE } from "app/apis/catalog/hca-atlas-tracker/common/entities";
 import { taskCompletionDatesSchema } from "app/apis/catalog/hca-atlas-tracker/common/schema";
-import { dbValidationToApiValidation } from "app/apis/catalog/hca-atlas-tracker/common/utils";
 import { METHOD } from "app/common/entities";
 import { updateTargetCompletions } from "app/services/validations";
 import { handler, method, role } from "app/utils/api-handler";
