+++ pages/api/users.ts
@@ -1,6 +1,6 @@
+import { dbUserToApiUser } from "../../app/apis/catalog/hca-atlas-tracker/common/backend-utils";
 import { ROLE_GROUP } from "../../app/apis/catalog/hca-atlas-tracker/common/constants";
 import { HCAAtlasTrackerDBUserWithAssociatedResources } from "../../app/apis/catalog/hca-atlas-tracker/common/entities";
-import { dbUserToApiUser } from "../../app/apis/catalog/hca-atlas-tracker/common/utils";
 import { METHOD } from "../../app/common/entities";
 import { getAllUsers, getUserByEmail } from "../../app/services/users";
 import {
