+++ pages/api/users/create.ts
@@ -1,6 +1,6 @@
+import { dbUserToApiUser } from "../../../app/apis/catalog/hca-atlas-tracker/common/backend-utils";
 import { ROLE } from "../../../app/apis/catalog/hca-atlas-tracker/common/entities";
 import { newUserSchema } from "../../../app/apis/catalog/hca-atlas-tracker/common/schema";
-import { dbUserToApiUser } from "../../../app/apis/catalog/hca-atlas-tracker/common/utils";
 import { METHOD } from "../../../app/common/entities";
 import { createUser } from "../../../app/services/users";
 import { handler, method, role } from "../../../app/utils/api-handler";
